+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.3.0] - 2021-09-10
+
 ### Added
 
 -   Add optional `--hud` CLI flag to allow choosing between expanding HUD horizontally (default) or centering the HUD.
@@ -193,7 +195,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Initial release.
 
-[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v1.3.0...HEAD
+
+[v1.3.0]: https://github.com/nbusseneau/hephaistos/compare/v1.2.0...v1.3.0
 
 [v1.2.0]: https://github.com/nbusseneau/hephaistos/compare/v1.1.1...v1.2.0